+++ CHANGELOG.md
@@ -8,14 +8,16 @@ FEATURES:
 * **New Argument:** `rancher2_cluster.rke_config.ingress.https_port` - (Optional/Computed) HTTPS port for RKE Ingress (int)
 * **New Argument:** `rancher2_cluster.rke_config.ingress.network_mode` - (Optional/Computed) Network mode for RKE Ingress (string)
 * **New Argument:** `rancher2_cluster.rke_config.ingress.update_strategy` - (Optional) RKE ingress update strategy (list Maxitems: 1)
+* **New Argument:** `rancher2_cluster.rke2_config` - (Optional/Computed) The RKE2 configuration for `rke2` Clusters. Conflicts with `aks_config`, `eks_config`, `gke_config`, `oke_config`, `k3s_config` and `rke_config` (list maxitems:1)
 
 ENHANCEMENTS:
 
-
+* Updated `rancher2_catalog_v2` schema resource, defining conflict between `git_repo` and `url` arguments
 
 BUG FIXES:
 
 * Fix `rancher2_node_pool` resource, adding `forcenew` property to not updatable arguments
+* Fix `rancher2_catalog_v2` resource, just setting default `git_branch` value if `git_repo` is specified
 
 ## 1.13.0 (March 31, 2021)